Add separate typography package
