+++ .gitignore
@@ -2,3 +2,5 @@
 public
 package-lock.json
 node_modules
+dist
+coverage
