+++ package.json
@@ -10,6 +10,7 @@
   },
   "workspaces": [
     "theme-ui",
+    "typography",
     "docs"
   ]
 }
