--- theme-ui/.gitignore
@@ -1,2 +1,0 @@
-coverage
-dist
