+++ typography/.npmignore
@@ -1,0 +1,4 @@
+babel.config.js
+test
+coverage
+
