+++ typography/babel.config.js
@@ -1,6 +1,5 @@
 module.exports = {
   presets: [
     '@babel/env',
-    '@babel/react',
   ],
 }
