+++ typography/package.json
@@ -1,36 +1,26 @@
 {
-  "name": "theme-ui",
+  "name": "theme-ui-typography",
   "version": "0.0.14",
   "main": "dist/index.js",
   "module": "dist/index.esm.js",
   "author": "Brent Jackson <jxnblk@gmail.com>",
   "license": "MIT",
-  "repository": "system-ui/theme-ui",
   "scripts": {
     "prepare": "microbundle --no-compress",
     "watch": "microbundle watch --no-compress",
     "test": "jest"
   },
   "dependencies": {
-    "@emotion/core": "^10.0.0",
-    "@emotion/styled": "^10.0.0",
-    "@mdx-js/react": "^1.0.0",
-    "@styled-system/css": "^2.0.2",
-    "emotion-theming": "^10.0.0",
-    "lodash.get": "^4.4.0",
-    "lodash.merge": "^4.6.0",
-    "react": "^16.8.0",
-    "styled-system": "^4.0.0"
+    "css-what": "^2.1.3",
+    "lodash.merge": "^4.6.1",
+    "typography": "^0.16.19"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "@mdx-js/mdx": "^1.0.0",
-    "jest": "^24.7.0",
-    "jest-emotion": "^10.0.10",
+    "@babel/core": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
+    "jest": "^24.8.0",
     "microbundle": "^0.11.0",
-    "react-test-renderer": "^16.8.6",
-    "react-testing-library": "^7.0.0"
+    "typography-theme-wordpress-2016": "^0.16.19"
   },
   "jest": {
     "testRegex": "./test/.*.js$",
